[박현석] - 게시물 삭제 기능 - 모달 활용
hspark9781/Hoicegram#14

+++ src/main/java/com/hsp/hoicegram/post/bo/PostBO.java
@@ -90,13 +90,19 @@ public class PostBO {
 	
 	public int deletePost(int userId, int postId) {
 		Post post = postDAO.selectPost(postId, userId);
-			
+		
+		if(post != null) {
 			FileManagerService.removeFile(post.getImagePath());
+			
+			likeBO.deleteLikeByPostId(postId);
+			commentBO.deleteCommentByPostId(postId);
+			
+			return postDAO.deletePost(userId, postId);
+			
+		} else {
+			return 0;
+		}
 		
-		likeBO.deleteLikeByPostId(postId);
-		commentBO.deleteCommentByPostId(postId);
-		
-		return postDAO.deletePost(userId, postId);
 	}
